Extract shared price encoder
